RestApi/api/requirements.txt
fastapi
uvicorn
pydantic
grpcio
grpcio-tools